user param 'time' fix

+++ model/model.go
@@ -18,7 +18,7 @@ type (
 		Name   string   `json:"name"`
 		RoomID string   `json:"room_id"`
 		Color  string   `json:"color"`
-		Time   int64    `json:"time"`
+		Time   int      `json:"time"`
 		Conn   net.Conn `json:"-"`
 	}
 )
